requirements.txt
dependency-injector==4.41.0
importlib-metadata==6.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mockfs==1.1.4
PyYAML==6.0.1
six==1.16.0
zipp==3.17.0
